Split extension support from hwcaps module
